transition remaining classes
