+++ src/core/Stocks/BuyTransaction.cs
@@ -1,0 +1,4 @@
+namespace core.Stocks
+{
+    public class BuyTransaction : StockTransaction {}
+}
