+++ src/core/Stocks/SellTransaction.cs
@@ -1,0 +1,4 @@
+namespace core.Stocks
+{
+    public class SellTransaction : StockTransaction {}
+}
