+++ src/core/Stocks/StockTransaction.cs
@@ -1,9 +1,10 @@
 using System;
 using System.ComponentModel.DataAnnotations;
+using MediatR;
 
 namespace core.Stocks
 {
-    public class StockTransaction
+    public class StockTransaction : IRequest
     {
         [Required]
         public string Ticker { get; set; }
@@ -16,5 +17,11 @@ namespace core.Stocks
         
         [Required]
         public DateTime? Date { get; set; }
+
+        public string UserId { get; private set; }
+        public void WithUser(string userId)
+        {
+            this.UserId = userId;
+        }
     }
 }
