+++ src/web/Controllers/PortfolioController.cs
@@ -2,6 +2,8 @@ using System.Linq;
 using System.Threading.Tasks;
 using core;
 using core.Options;
+using core.Portfolio;
+using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using web.Utils;
@@ -13,46 +15,19 @@ namespace web.Controllers
     [Route("api/[controller]")]
     public class PortfolioController : ControllerBase
     {
-        private IPortfolioStorage _storage;
+        private IMediator _mediator;
 
-        public PortfolioController(IPortfolioStorage storage)
+        public PortfolioController(IMediator mediator)
         {
-            this._storage = storage;
+            this._mediator = mediator;
         }
 
         [HttpGet]
         public async Task<object> Index()
         {
-            var stocks = await _storage.GetStocks(this.User.Identifier());
+            var query = new Get.Query(this.User.Identifier());
 
-            var cashedout = stocks.Where(s => s.State.Owned == 0);
-            var owned = stocks.Where(s => s.State.Owned > 0);
-
-            var totalSpent = stocks.Sum(s => s.State.Spent);
-            var totalEarned = stocks.Sum(s => s.State.Earned);
-
-            var options = await _storage.GetSoldOptions(this.User.Identifier());
-
-            var ownedOptions = options.Where(o => o.State.Amount > 0);
-            var closedOptions = options.Where(o => o.State.Amount == 0);
-            
-            var obj = new
-            {
-                totalSpent,
-                totalEarned,
-                totalCashedOutSpend = cashedout.Sum(s => s.State.Spent),
-                totalCashedOutEarnings = cashedout.Sum(s => s.State.Earned),
-                owned = owned.Select(o => StocksController.ToOwnedView(o)),
-                cashedOut = cashedout.Select(o => StocksController.ToOwnedView(o)),
-                ownedOptions = ownedOptions.Select(o => GetSoldOption.Handler.ToOptionView(o)),
-                closedOptions = closedOptions.Select(o => GetSoldOption.Handler.ToOptionView(o)),
-                pendingPremium = ownedOptions.Sum(o => o.State.Premium),
-                collateralCash = ownedOptions.Sum(o => o.State.CollateralCash),
-                collateralShares = ownedOptions.Sum(o => o.State.CollateralShares),
-                optionEarnings = options.Sum(o => o.State.Profit)
-            };
-
-            return obj;
+            return await _mediator.Send(query);
         }
     }
 }
