+++ src/web/Controllers/StocksController.cs
@@ -4,6 +4,7 @@ using System.Threading.Tasks;
 using core;
 using core.Adapters.Stocks;
 using core.Stocks;
+using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using web.Utils;
@@ -15,125 +16,35 @@ namespace web.Controllers
     [Route("api/[controller]")]
     public class StocksController : ControllerBase
     {
-        private IStocksService _stocksService;
-        private IPortfolioStorage _storage;
+        private IMediator _mediator;
 
-        public StocksController(
-            IStocksService stockService,
-            IPortfolioStorage storage)
+        public StocksController(IMediator mediator)
         {
-            _stocksService = stockService;
-            _storage = storage;
+            _mediator = mediator;
         }
 
         [HttpGet("{ticker}")]
         public async Task<object> DetailsAsync(string ticker)
         {
-            var profile = await _stocksService.GetCompanyProfile(ticker);
-
-            var data = await _stocksService.GetHistoricalDataAsync(ticker);
-
-            var price = data.Historical.Last().Close;
-
-            var largestGain = data.Historical.Max(p => p.ChangePercent);
-            var largestLoss = data.Historical.Min(p => p.ChangePercent);
-
-            var byMonth = data.Historical.GroupBy(r => r.Date.ToString("yyyy-MM-01"))
-                .Select(g => new
-                {
-                    Date = DateTime.Parse(g.Key),
-                    Price = g.Average(p => p.Close),
-                    Volume = g.Average(p => p.Volume),
-                    Low = g.Min(p => p.Close),
-                    High = g.Max(p => p.Close)
-                });
-
-            var labels = byMonth.Select(a => a.Date.ToString("MMMM"));
-            var lowValues = byMonth.Select(a => Math.Round(a.Low, 2));
-            var highValues = byMonth.Select(a => Math.Round(a.High, 2));
-
-            var ratings = await _stocksService.GetRatings(ticker);
-            var ratingInfo = new
-            {
-                rating = ratings.Rating?.ToString() ?? "not available",
-                details = ratings.RatingDetails?.Select(d => new
-                {
-                    name = d.Key,
-                    rating = d.Value
-                })
-            };
-
-            var priceValues = byMonth.Select(a => Math.Round(a.Price, 2));
-            var priceChartData = labels.Zip(priceValues, (l, p) => new object[] { l, p });
-
-            var volumeValues = byMonth.Select(a => a.Volume);
-            var volumeChartData = labels.Zip(volumeValues, (l, p) => new object[] { l, p });
-
-            var metrics = await _stocksService.GetKeyMetrics(ticker);
-            var mostRecent = metrics.Metrics.FirstOrDefault();
-
-            int age = 0;
-            if (mostRecent != null)
-            {
-                age = (int)(DateTime.UtcNow.Subtract(mostRecent.Date).TotalDays / 30);
-            }
-
-            var metricDates = metrics.Metrics.Select(m => m.Date.ToString("MM/yy")).Reverse();
-
-            var bookValues = metrics.Metrics.Select(m => m.BookValuePerShare).Reverse();
-            var bookChartData = metricDates.Zip(bookValues, (l, p) => new object[] { l, p });
-
-            var peValues = metrics.Metrics.Select(m => m.PERatio).Reverse();
-            var peChartData = metricDates.Zip(peValues, (l, p) => new object[] { l, p });
-
-            return new
-            {
-                ticker,
-                price,
-                profile = profile.Profile,
-                age,
-                bookValue = mostRecent?.BookValuePerShare,
-                peValue = mostRecent?.PERatio,
-                ratings = ratingInfo,
-                largestGain,
-                largestLoss,
-                labels,
-                priceChartData,
-                volumeChartData,
-                bookChartData,
-                peChartData
-            };
+            return await _mediator.Send(new Get.Query(ticker));
         }
 
         [HttpPost("sell")]
-        public async Task<ActionResult> Sell(StockTransaction model)
+        public async Task<ActionResult> Sell(SellTransaction model)
         {
-            var stock = await this._storage.GetStock(model.Ticker, this.User.Identifier());
-            if (stock == null)
-            {
-                return NotFound();
-            }
+            model.WithUser(this.User.Identifier());
 
-            stock.Sell(model.Amount, model.Price, model.Date.Value);
-
-            await this._storage.Save(stock);
+            await _mediator.Send(model);
 
             return Ok();
         }
 
         [HttpPost("purchase")]
-        public async Task<ActionResult> Purchase(StockTransaction model)
+        public async Task<ActionResult> Purchase(BuyTransaction model)
         {
-            var stock = await this._storage.GetStock(model.Ticker, this.User.Identifier());
+            model.WithUser(this.User.Identifier());
 
-            if (stock == null)
-            {
-                stock = new OwnedStock(model.Ticker, this.User.Identifier());
-            }
-
-            stock.Purchase(model.Amount, model.Price, model.Date.Value);
-
-            await this._storage.Save(stock);
+            await _mediator.Send(model);
 
             return Ok();
         }
@@ -141,29 +52,16 @@ namespace web.Controllers
         [HttpGet("export")]
         public async Task<ActionResult> Export()
         {
-            var stocks = await _storage.GetStocks(this.User.Identifier());
-
-            var filename = CSVExport.GenerateFilename("stocks");
+            var response = await _mediator.Send(new Export.Query(this.User.Identifier()));
 
             this.HttpContext.Response.Headers.Add(
                 "content-disposition", 
-                $"attachment; filename={filename}");
+                $"attachment; filename={response.Filename}");
 
             return new ContentResult
             {
-                Content = CSVExport.Generate(stocks),
-                ContentType = "text/csv"
-            };
-        }
-
-        internal static object ToOwnedView(OwnedStock o)
-        {
-            return new
-            {
-                ticker = o.State.Ticker,
-                owned = o.State.Owned,
-                spent = Math.Round(o.State.Spent, 2),
-                earned = Math.Round(o.State.Earned, 2)
+                Content = response.Content,
+                ContentType = response.ContentType
             };
         }
     }
